more shitty fucking refactoring im lost in the sauce
