+++ src/layouts/Layout.astro
@@ -1,6 +1,6 @@
 ---
 import ThemeInit from "../styles/ThemeInit.astro";
-import { defaultTheme } from "../styles/themeData";
+import { defaultTheme } from "../styles/config";
 import ThemeSelector from "../styles/ThemeSelector.svelte";
 import baseStylesUrl from "../styles/base.css?url";
 import resetStylesUrl from "../styles/reset.css?url";
